BEOS_*.* files moved into src/ditectory GUI files compared with original version of this program. Only difference is usage of int64_t instead for off_t, no changes made. Littlebit cleanup SystemDependencies removed (it's useless)

+++ src/add-ons/kernel/file_systems/befs/BEFS_Inode.h
@@ -277,7 +277,7 @@ class AttributeIterator {
 		
 		status_t Rewind();
 		//status_t GetNext(char *name, size_t *length, uint32_t *type, vnode_id *id);
-		status_t GetNext(char *name, size_t *length, uint32_t *type, unsigned long long *id);
+		status_t GetNext(char *name, size_t *length, uint32_t *type, uint64_t *id);
 
 	private:
 		friend class Chain<AttributeIterator>;
